# docs/files/swagger.yaml
openapi: 3.0.0
info:
  title: APIWeb - Online Marketplace API
  description: A RESTful API for an online marketplace platform
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
        image_url:
          type: string
        category_product_id:
          type: integer
        user_id:
          type: integer
    Category:
      type: object
      properties:
        id:
          type: integer
        category:
          type: string
paths:
...
  /products:
    get:
      tags:
        - Products
      summary: Get all products
      responses:
        200:
          description: List of products retrieved successfully
          
    post:
      tags:
        - Products
      summary: Create a new product
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        201:
          description: Product created successfully
...
  /categories:
    get:
      tags:
        - Categories
      summary: Get all categories
      responses:
        200:
          description: List of categories retrieved successfully
          
    post:
      tags:
        - Categories
      summary: Create a new category
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        201:
          description: Category created successfully
...